Apply design patterns make the design modular
To be able to work as it is growing, some design patterns were mandatory.
Under source now, there are all the _sources_ which can be used to the
generator. The abstraction allows us with:
- incremental feature additions without modifying core logic
- pluggable implementations of the same interface

The ContextBuilder implemented with the Facade Pattern in mind. This
hides complexity from LLM module and provides a clean separation of
concerns. it needs to know only one method rather than handling each
source separately.

With the above changes, the target is the future land of implementations,
getting ready to adapt and extend the scope of the usability and
integrate other ideas, as VectorDB for the datasets.

Additional, there were improvements in the cli to simplify the usage via
two commands. At the current state is only ready for a demo purposes.
It lacks checks and real environment setup to run on. In theory this
should be some container or VM. For now it runs against the local env.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>